requirements_dev.txt
pytest-django>=3.7.0
pytest-asyncio>=0.10.0